electric_indigo @ 20% tint

One of the five shades of electric_indigo. The number is a percent of the original color, ranging from 100 (100%) down to 20 (20%) at 20% intervals. The lower the number, the lighter the tint of the original color.

Usage

electric_indigo_20

Format

An object of class character of length 1.

See also